Robust NuGet builds
Building Scintilla.NET now forces Scintilla.NET.TestApp to rebuild by deleting
`Scintilla.NET.TestApp\obj\*.nuget.*` files. This makes it possible to properly
test NuGet build by opening `Scintilla.NET-Nuget.sln` and
`Scintilla.NET.TestApp-Nuget.sln` side by side.

+++ Scintilla.NET/nupkg.ps1
@@ -1,26 +1,36 @@
-param (
-	[Parameter(Mandatory=$true)][string]$MSBuildProjectDirectory,
-	[Parameter(Mandatory=$true)][string]$NuGetPackageRoot,
-	[Parameter(Mandatory=$true)][string]$NuGetPackageSourceDir,
-	[Parameter(Mandatory=$true)][string]$PackageOutputPath,
-	[Parameter(Mandatory=$true)][string]$PackageId,
-	[Parameter(Mandatory=$true)][string]$PackageVersion
+﻿param (
+    [Parameter(Mandatory, ParameterSetName="Clean")][switch]$Clean,
+    [Parameter(Mandatory, ParameterSetName="Add")][switch]$Add,
+    [Parameter(Mandatory)][string]$MSBuildProjectDirectory,
+    [Parameter(Mandatory, ParameterSetName="Clean")][string]$NuGetPackageRoot,
+    [Parameter(Mandatory)][string]$SolutionDir,
+    [Parameter(Mandatory)][string]$PackageOutputPath,
+    [Parameter(Mandatory)][string]$PackageId,
+    [Parameter(Mandatory)][string]$PackageVersion
 )
 
 Set-Location -Path $MSBuildProjectDirectory
 
-$PackageVersionList = $PackageVersion.Split('.')
+$NuGetPackageSourceDir = Join-Path $SolutionDir "registry"
 
-if ($PackageVersionList.Count -Eq 4 -And [int]::Parse($PackageVersionList[3]) -Eq 0) {
-	$PackageVersion = [string]::Join('.', $PackageVersionList[0..2])
+if ($Clean) {
+    $PackageInRegistry = Join-Path $NuGetPackageSourceDir $PackageId
+    $PackageInRoot = Join-Path $NuGetPackageRoot $PackageId
+    $TestAppObjDir = Join-Path $SolutionDir "Scintilla.NET.TestApp\obj"
+    
+    Remove-Item $PackageInRegistry -Recurse -Force -ErrorAction Ignore
+    Remove-Item $PackageInRoot -Recurse -Force -ErrorAction Ignore
+    Get-ChildItem $TestAppObjDir -Filter "*.nuget.*" -ErrorAction Ignore | Remove-Item -Force -ErrorAction Ignore
 }
 
-$PackagePath = Join-Path $PackageOutputPath "$PackageId.$PackageVersion.nupkg"
-
-$PackageInRegistry = Join-Path $NuGetPackageSourceDir $PackageId
-$PackageInRoot = Join-Path $NuGetPackageRoot $PackageId
-
-Remove-Item $PackageInRegistry -Recurse -Force -ErrorAction Ignore
-Remove-Item $PackageInRoot -Recurse -Force -ErrorAction Ignore
-
-nuget add $PackagePath -Source $NuGetPackageSourceDir
+if ($Add) {
+    $PackageVersionList = $PackageVersion.Split('.')
+    
+    if ($PackageVersionList.Count -Eq 4 -And [int]::Parse($PackageVersionList[3]) -Eq 0) {
+        $PackageVersion = [string]::Join('.', $PackageVersionList[0..2])
+    }
+    
+    $PackagePath = Join-Path $PackageOutputPath "$PackageId.$PackageVersion.nupkg"
+    
+    nuget add $PackagePath -Source $NuGetPackageSourceDir
+}
